refactor: move themes into board addon crates, feature-gate esp-hal
PaperUI is now strictly the declarative plumbing — engine, widget logic, and the
WidgetTheme *contract*. Concrete looks move to the board addons that own rendering:
  - DefaultTheme -> paperui-tft::TftTheme
  - EinkTheme    -> paperui-eink::EinkTheme

Each board crate gains a default-on `hal` feature gating esp-hal (ButtonReader /
GT911 touch); the theme is pure logic over Canvas and stays host-testable via
`--no-default-features --features mock`. The IT8951 driver/renderer/canvas are
generic over embedded-hal traits and now compile on host too.

Coverage unchanged: paperui 17 + paperui-tft 3 + paperui-eink 3 = 23 host tests,
all green; both board crates build for xtensa (default features); clippy clean on
host + device (fixed 2 nits surfaced by host-linting the driver).

+++ crates/paperui-eink/src/canvas.rs
@@ -35,9 +35,9 @@ impl Default for Gray4Canvas {
 }
 
 fn gray4(c: Color) -> u8 {
-    let r = ((c.0 >> 16) & 0xFF) as u32;
-    let g = ((c.0 >> 8) & 0xFF) as u32;
-    let b = (c.0 & 0xFF) as u32;
+    let r = (c.0 >> 16) & 0xFF;
+    let g = (c.0 >> 8) & 0xFF;
+    let b = c.0 & 0xFF;
     let luma = (r * 54 + g * 183 + b * 19) >> 8;
     (luma >> 4) as u8
 }
